+++ derrida/books/views.py
@@ -9,12 +9,11 @@ from django.db.models import Max, Min
 from django.http import HttpResponseRedirect, HttpResponse, Http404
 from django.shortcuts import get_object_or_404
 from django.urls import reverse
-from django.utils import timezone
 from django.utils.decorators import method_decorator
 from django.views.generic import DetailView, ListView, View
 from django.views.generic.base import TemplateView
 from django.views.generic.edit import FormView
-from djiffy.models import Canvas
+from djiffy.models import Canvas, get_iiif_url
 from haystack.query import SearchQuerySet
 from haystack.inputs import Clean, Raw
 import requests
@@ -58,6 +57,22 @@ class InstanceDetailView(DetailView):
     def get_queryset(self):
         instances = super(InstanceDetailView, self).get_queryset()
         return instances.filter(digital_edition__isnull=False)
+
+    def get_context_data(self, *args, **kwargs):
+        '''Insert manifest license data into template context.'''
+        context_data = super(InstanceDetailView, self).get_context_data()
+        instance = context_data['object']
+        license_url = instance.digital_edition.license
+        if license_url:
+            data_url = license_url.split('?')[0].replace('vocab', 'data')
+            response = requests.get(data_url)
+            if response.status_code == 200:
+                # This should be safe for all rightsstatments.org
+                # prefLabels as they seem to be setting English as their
+                # default using @ notation.
+                context_data['license_text'] = \
+                    response.json()['prefLabel']['@value']
+        return context_data
 
 
 class InstanceReferenceDetailView(InstanceDetailView):
@@ -499,10 +514,20 @@ class CanvasDetail(DetailView):
         Set extra context for :class:`CanvasDetail` view.
         '''
         context = super(CanvasDetail, self).get_context_data(*args, **kwargs)
+        # If there is a plain_text_url in info, use a Djiffy method to get the
+        # text from Figgy and pass it to the view, default ocr_text to None
+        ocr_text = None
+        if self.object.plain_text_url:
+            # get the text
+            res = get_iiif_url(self.object.plain_text_url)
+            # check that we got a valid response and set ocr_text if so.
+            if res.status_code == 200:
+                ocr_text = res.text
         context.update({
             'instance': self.instance,
             'canvas_suppressed': self.instance.suppress_all_images or \
-                    self.object in self.instance.suppressed_images.all()
+                    self.object in self.instance.suppressed_images.all(),
+            'ocr_text': ocr_text
         })
         if self.request.user.has_perm('books.change_instance'):
             context.update({